Remove build-time Git error flood
The original Makefile tries to calculate a Git revision number from the
source tree. Unless the source tree happens to be a checked-out
Git worktree, that causes standard output to be flooded with lots of
unhelpful error messages and help texts.
If Git is not installed at build time, similar error messages would
occur.

Short-circuit the Git revision calculation to prevent this.