ksh93.install: Remove false assumption regarding /usr/bin
Just because an entry in /etc/shells exists for /bin/*ksh doesn't mean
an equivalent entry for /usr/bin also exists in the file. This could
cause the following error to occur in some rare situations:

The value for the SHELL variable was not found in the /etc/shells file

A for loop is now used to explicitly check for both entries, which
should mitigate that rare issue.