Update to 5.264.1
Native package is at 5.229.1. However, that version is broken against
latest vulkan-headers. Since the native package has already been flagged
as out of date, I'm updating to the latest version and this update should
be available once the native package is also updated.

Merge in the changes from the native package that had were added since
4.208.0

Signed-off-by: Alexandre Demers <alexandre.f.demers@gmail.com>