Make the package be more isolated and include queries
Switching from -uNORC to -uNONE means all files in the runtimepath will
be ignored and not automatically loaded at startup, nearly completely
isolating the neovim instance from the system. It shouldn't be impacted
by anything installed elsewhere.

The queries shipped in the repo are meant to match with the version of
the parser marked in the lock file, so ship the queries as well.
Otherwise a mismatch between the parser and query versions will lead to
a bunch of treesitter errors, especially in the highlighter.