Install library in dedicated namespace
Since adding encryption support changes the ABI, it is not feasible to
simply replace the official package's files and require users to
recompile all downstream software.

To work around that, I have modified the build system of this package
to install everything as `QuotientE2EE` instead of `Quotient`. So the
headers are installed to `/usr/include/QuotientE2EE` instead of
`/usr/include/Quotient`, the library itself is installed as
`libQuotientE2EE.so` instead of `libQuotient.so` and the `pkg-config`
and CMake configuration files have been renamed and adjusted
accordingly to allow compilation against this separate library.

As a side-effect of this, it should be possible to have `libquotient`
and `libquotient-encryption` installed at the same time - there are no
conflicting files between these packages now.

The preprocessor macros passed to the compiler are unchanged, as are
the CMake configuration options passed during the build process. At
least in theory, all that needs to change in downstream software is
to make the downstream build system use the `QuotientE2EE` library
instead of `Quotient` and to include headers from the `QuotientE2EE`
system directory instead of `Quotient` - all other variables should be
able to remain unchanged.

The header files are an interesting case, as they expect to be
installed in a directory named `Quotient`. This is something I don't
think I can change in the build system itself without moving the entire
source tree or separating the headers from the source, so I'll probably
have to leave the headers as-is during the build, and then change them
in the `package()` step. All that should need to change in the headers
is replacing instances of `#include <Quotient` with
`#include <QuotientE2EE` - everything else should continue to work.