Patch headers to support new install location
After CMake has finished installing the header files in the package
directory, a pass is run over all of them to replace all instances of
`# include <Quotient/` with `# include <QuotientE2EE/` to ensure that
the nested header includes work against the new install location. The
regular expression I used is sufficiently flexible to support a variety
of white-space scenarios and should be future-proofed for the inclusion
of new headers in the future.