Added patch to fix bug with viewing images
This is due to Webkit's hardware accelerated compositor frankly sucking,
or OpenGL (could be either).

When resizing to view an image on a post with multiple images, Webkit
ends up taking an extremely long time to load the rest of the UI. Why, I
have no idea, but it inhibits usability so it was best to disable
the hardware accelerated compositor.