Update patchset for upstream 0.8.0 release
As of 0.8.0, libQuotient has a build-time option for selecting either
Qt 5 or Qt 6, which required adding infrastructure to change the name
of the library target to prevent conflicts. Since I had to do this in
my patchset for 0.7.2, the patchset is now greatly simplified.

I still needed to make a few changes to other files to move the header
install location though. I have also made some changes to the
`pkg-config` description file to make it more responsive to the
selected build configuration. These changes should probably be
upstreamed as the file is otherwise incorrect for Qt 6 builds.