Update to upstream version 23.04.3
The patch needed to make NeoChat compatible with libQuotient 0.8
didn't make it into this release. The official `neochat` package on
Arch Linux manually applies this patch so that the `libquotient`
package can also stay up to date, so I do the same here, on top of the
other patches I wrote to make the build use the
`libquotient-encryption` package instead.