Update build process and add pandoc to makedepends
Since the setup.py build process is deprecated, switch to the
standards-based one [1] – seems to work like a charm.

Since the new build includes a pandoc step and fails if pandoc isn’t
available, add that to the makedepends. Everyone who wants to use this
package will just have to live with the massive package list, disk space
usage, and slew of upgrades that installing pandoc brings with it.

[1]: https://wiki.archlinux.org/index.php?title=Python_package_guidelines&oldid=781620#Standards_based_(PEP_517)