remove example service file and install script
Installing the service file to /usr/lib/systemd/system made little sense
anyways as it can't be used without editing it anyways. Also, by far not the
only way to run Trac, I trust users to be able to pick their preferred method
and implement it for their setup themselves. The install script became obsolete
as it basically just said that the installed service file isn't usable in the
default state.