Change package name without -git, .SRCINFO added
This PKGBUILD does not follow the latest commit from the upstream repo
and it having `-git` is an old artifact I neglected to remove. This
makes this repo a bit more compliant for the AUR. There might be a
future branch/repo that actually uses the latest commit, however.

`.SRCINFO` was also added in preperation for AUR publishing.