0.5.9 -> 0.5.13
We had to skip a few versions because of release issues:
<https://github.com/lidatong/dataclasses-json/issues/438#issuecomment-1638736629>

This module is now using poetry instead of setuptools/pipenv:
<https://github.com/lidatong/dataclasses-json/commit/094f0742faa2>

We had to remove the .gitignore file in this repo because it was preventing the
poetry package from building the wheel properly.  I believe this issue is
related: <https://github.com/python-poetry/poetry/issues/2809>

I also decided to completely overhaul update.sh since it was getting unwieldy
while testing the massive PKGBUILD updates.