Update to 2.1.11
Additionally, let's add Apache to the list of licenses. It looks like
the authors do not know how the code is licensed and also do not care as
they do not plan to clarify this. Best guess is that it is dual licensed
under GPL3 and Apache.

See https://github.com/meshtastic/python/issues/422