Install the thumbnailer programs to /usr/bin instead of /usr/local/bin
The original /usr/local/bin path was copy-pasted from the upstream
project's README file's installation instructions.  However, that path
is only for software that user's install themselves, circumventing the
package manager, and because this is actually packaging the software,
that's not the right place to put it.  It really belongs in /usr/bin
where all the other package-managed programs go.