Fix and always run hypothesis based tests
If `python2-hypothesis` happens to be installed, the test suite run additional tests. Those always
fail with an import error unless `python2-mock` also happens to be installed.

If both are present, one of the tests will frequently but not always cause hypothesis to error out
complaining about too much of the generated input being filtered out or the test running too slow.
Those should really be warnings and not cause tests to fail but here we are...

- Add `python2-hypothesis` and `python2-mock` to checkdepends so the tests run everywhere.
- Silence the health check errors. If there is a way to do that from the command line instead of
  patching the source, I didn't find it.