fix package(): do not touch real /opt/freecalypso
While being executed by makepkg, functions in PKGBUILD have access
to whatever the build-user has access to.  The current implementation
of package() would mess up with the real '/opt/freecalypso' directory.

The right approach is to override the INSTALL_PREFIX variable when
invoking make install, so that everything is installed directly
to the ${pkgdir}.