update to 1.2.4.3
git submodules have been removed, they're all installed via FetchContent
now. Thus, the prepare() and build() functions are adapted to reflect
this change.

Furthermore, since the repo doesn't contain any git submodules anymore,
the source is changed to the release .tar.gz archives, also for
googletest and libbpf. Since p4runtime doesn't have any recent releases,
the git repository is still used there.

libbpf needs to be present in a certain directory for the build to
succeed, so it is symlinked in the prepare() function.

The protobuf 23 patch has been accepted upstream and is removed as a
separate source.