update to 2.0.0.0
Note that the patch regarding VARINT static assert failure:
the argument is unsignd, so the DEFAULT VarIntMode works for
it. So, we fixed the failure by removing the mode, letting
it be default. But, upstream did the other fix, which is to cast to
arg to a signed type, keeping the non-default VarIntMode.

    b1b51490058bbc693eca8ef7abd69dc022446201
    cast nHeight to int32_t to match the VarIntMode expected type/value

https://gitlab.com/bitcoinunlimited/BCHUnlimited/-/commit/b1b51490058bbc693eca8ef7abd69dc022446201