Enable native support for Wayland
By default the Electron (Chromium Engine) uses XWayland adding a
unnecessary layer. I added `--ozone-platform-hint=auto` in the
tetrio-launcher.sh script so if the user is on X11, it will choose the
X11 "Ozone platform" and if the user is on Wayland it will try to use
the Wayland "Ozone platform".