Update to 22.1.7 and tweak flags
disabled glx, enabled lto, disabled gles1
These flags seem more in line with what meta-riscv are using (afaict)
and I was successfully able to get a few sdl apps and retroarch to
definitively use the driver rather than falling back to swrast
or crashing like they were previosuly.

Unsure if this breaks kde plasma or if it was broken by the upstream
driver investigation I was doing which involved copying random shit to /usr/lib
XD. Regardless, it was sw rendering before as far as I can tell so not
much of a loss.