Update provides list
Context: https://aur.archlinux.org/packages/bitwig-studio#comment-933914

+1 for updating the provides array: provides=(clap-host vst-host vst3-host).
This is in line with the current packaging practices for audio software
in Arch. Having only Bitwig and a few packaged vst plugins means that
another host has to be installed to satisfy the *-host dependencies. Ex:
https://archlinux.org/packages/extra/x86_64/dexed-vst3/

It is also not necessary anymore to have bitwig-studio listed:
https://wiki.archlinux.org/title/Arch_package_guidelines#Package_relations

"Do not add $pkgname to PKGBUILD#provides, as it is always implicitly
provided by the package."