Bring up the nextdns service before the network-online.target
Since some services like Wireguard depend on the network-online.target
make sure nextdns is brought up before that,
otherwise Wireguard won't be able to resolve the hostnames and fail to start.

Nextdns is more forgiving and can connect to its endpoints at a later time,
when the network is online.