Version bump to v13.0
The v13.0 release has a number of changes in the filenames that are
generated. I've modified the `_dist_checksum()` fn to match the new
filenames as well as removed the suffix which is no longer present.
The `_tag_i686`/`tag_x86_64` could likely be simplified to use the
`_archstr`, rather than hardcoding.