Update to 5.11
csl-styles and csl-locales are now included using git submodules and
they are not included in the tarball anymore. Using git for building
this package would require cloning hundreds of MB of repository data.
Hence, we now manually download tarballs for all submodules and need to
maintain a list of correct commit hashes.

Also, let's start using Arch's gradle package again as it is currently
capable of building JabRef.