Support New portals req for XDG_CURRENT_DESKTOP
As of version 1.17, xdg-desktop-portal requires desktops to declare
which portals should be used. For Sway, this requires a file
/usr/share/xdg-desktop-portal/sway-portals.conf (or a local user
config).

A reasonable default is to use GTK for the file picker (as wlr doesn't
have file picker support) and wlr for screencast/screenshot support.
Fedora does this similarly. [1]

Upstream is not interested in shipping a configuration file [2] so we
need to do it downstream.

Fixes #72297
Fixes #79883

[1] https://src.fedoraproject.org/rpms/sway/blob/rawhide/f/sway-portals.conf
[2] https://github.com/swaywm/sway/issues/7773