pkgbuild: re-add provides() for ceph-libs
As there are apparently people depending on ceph-libs, for which this
package _can_ be used.

This isn't ideal, as the other ceph-libs is a virtual package, so the
behavior between this one and the other will conflict in interesting
ways if I'm not careful.

Need to think of a better solution long term...

Reported-by: https://aur.archlinux.org/account/xeruf
References: https://aur.archlinux.org/pkgbase/ceph-bin#comment-940725