xdg-utils-mimeo: cover default-url-scheme-handler
Noticed a report regarding this missing and looking into what needs to be done it seemed fairly straighforward.
Only slightly tested and it seemed to be working. We will see what the real world says about this...

Funny sidenote: Last update to the patch was on the 9th nov of 2022 :D