bump version, allow building on x86_64
the mpp library itself works just fine on x86_64, software side, albeit without
actual hardware support, this should make some pkgs that wants mpp on x86_64
happy, and also makes it easier for us to test the build. it is then possible
to completely replace rockchip-mpp with this package, functionality-wise.