update: v6.58, new maintrainer
My attempt to revive this package and make it work.
The upstram project author focuses mainly on binary
distribution.

The biggest problem imho in packing this is storing credentials
in users config dir and NOT in gam.py folder. Previous maintaner
tried to bring the issue (https://github.com/GAM-team/GAM/issues/792),
but nobody seems to be bothered to do something about it.

The xdg_config_dirs.patch works good enough for me to try
maintaning this package.