Rebuild for gawk 5.3.0 (API major version bump)
Note that as a consequence of the gawk API bump, gawkextlib must be built and
installed before this package is rebuilt.

If you are using an AUR helper tool that does not build/install updated AUR
packages sequentially (in dependency order), then update to gawkextlib 1.0.4-4
first before updating this package.