Hardest decision in my life Made following changes:        deleted:    fix-arch-linux-desktop-portal.patch        deleted:    qt_scale_factor-fix.patch_        deleted:    workaround-for-dbusactivatable.patch        bumped:     pkgrel Why? Apparently someone should mention on ArchWiki that those variables are plague and will break the UX when e.g. clicking on a PDF file in tdesktop and okular launches. Also 0.8 is apparently hardcoded for the 1.25 case (1.25 * 0.8 = 1) while when e.g. the user has 1.5 scale, it will break
QT_QPA_PLATFORMTHEME=xdgdesktopportal would also break as soon as the
user prefers e.g. qt6ct

Qt doesn't acknowledge xdg-desktop-portal as something to be used
outside of flatpak/snap so there's no way to use the portals while
being able to specify what Qt should fallback to when QPT is being
queried for something portals don't handle

so it fallback to the hardcoded list that uses gtk3 QPT on GNOME, kde
QPT on KDE and QGenericUnixTheme on everything else (which is
terrible)

In short, that patches aren't needed and isn't used in other Telegram
clients. (Only AyuGram and materialgram have those patches)