Fix building with a custom SRCDEST
When building with makepkg with a SRCDEST environment variable
with nested directories (e.g. 'a/b/c') gendesk is unable to
find the PKGBUILD and generates a wrong .desktop file.

Explicitly pass --pkgname and --pkgdesc (which seems to be
what most official Arch packages do) to fix this issue.

Also get rid of the setconf makedepends by using --categories.