Replace broken download link with archive.org
Looks like darcs.idyll.org is no longer configured in DNS, along with
idyll.org. But http://ivory.idyll.org/ is working and I couldn't get a
read on how long the domains have been offline, so I don't know if this
is temporary or permanant. Certainly, permanent is a safe assumption
giving the age of the software. The link can be replaced back with the
original if it comes back online.