Make sure to launch JabRef with installed JDK
Arch Linux's official JDK and JRE packages of the same version conflict
with each other. Hence, depending on java-environment=21 works in the
majority of cases. However, there are JDK packages in the AUR that
provide java-environment but not jdk-openjdk and hence do not conflict
with JRE packages. Make sure to launch JabRef with a full JDK and don't
rely on it being chosen automatically.

Fixes a1fe8de0578a5e0e791a10af02d0cbd594dd3374