fix build (don't build all features), adjust for conduit-matrix name
`conduit` and `matrix-conduit` files conflict with the
matrix-conduit-bin AUR package and conduit-bin AUR package.

the build has been failing due to compiling all features which is not
recommended for Conduit as some database formats are not working
properly. ideally the project should remove or fix the broken code, but
that doesn't appear to be happening any time soon.
rocksdb is the only recommended database format either way, and sqlite
is included in the default features list for those that want to use it.

Signed-off-by: strawberry <strawberry@pupbrain.dev>