all: fix and update pkgver to tip
I believe there is something wrong with the latest tag, even thought it
exist in my local and github [1].

I think I may have re-pushed the tag with different commit hash earlier
but use the old hash in AUR which cause this issue.

The fix is by adding new commit and update the pkgver.

[1]: https://github.com/shuLhan/rescached-go/releases/tag/v4.4.2