Publish version 2023.12.04.00-2 - Build shared object
Build a shared objects (.so files) instead of static library (.a files).
This is more common place in line in Arch Linux packages and has the
following consequences:

* Downstream packages must specify folly in depends rather than
  makedepends.

* Smaller installed size (13.5 MiB vs 25.5 MiB).

Also, introduce testing. This makes the build significantly slower, 16
minutes in clean chroot on my system. This can however be mitigated by
using ccache, building in a tmpfs and by setting MAKEFLAGS="-j$(nproc)".