update to 13.5a2
this switches us back to the alpha channel since it is the latest version.  this necessitates a desktop file change which may require users to reset their dock icons. i waited a while as i looked for ways to avoid this, and have already missed the first two alpha releases. this can probably only be avoided with a custom compilation of the app, which is not acceptable for a binary package.

this also switches from the mullvad cdn to github. mullvad's cdn is painfully slow and not practical. they release everything on github so it works a lot better.