python-varname: Use PyPi prebuilt wheel
I've had a lot of trouble managing dependencies while building from
source, the `python-poetry-plugin-export` package is very slow to update
in the core repos, which this uses, and the dependencies of that package
update more quickly, so something or the other almost always breaks.

So, until this situation is resolved, just use the prebuilt wheel in
order to have a _functioning_ package.

Signed-off-by: Aseem Athale <athaleaseem@gmail.com>