Fix missing dependencies
This adds a number of dependencies from obs-studio-git that were
missing from this package (which was based on extra/obs-studio).

Importantly, asio was missing and prevented the build from working
at all, but several other plugins that obs-studio-git enabled
were not set up in this package. They should now be built as well.