Remove pickle import patch
See
https://github.com/ssb22/lexconvert/issues/5#issuecomment-1863015869,
trusting upstream to have placed enough assertions to gate the cPickle
import to only run under python2.
This python2/3 difference shouldn't be a problem -- as noted in the
link, python2 support is there primarily in order to support MacOS's
system voices. Documenting this using a python2 optdepend, even though
this is most likely irrelevant on Arch.