Fix binary paths in *_thumbnailer files
The previous fix installed the binaries to the semantically correct
place but the paths referencing them in the _thumbnailer files were not
updated.  So, although the binaries were present in the correct place
and ran without problem when called with the same arguments, the
meta-data in the _thumbnailer definitions was pointing to the wrong
place causing thumbnail factory failures when trying to call the
non-existing binaries.  Beginner mistake, whoops.

This fix removes the "local/" path part from all the thumbnailer files
during the prepare stage, before they are installed.