0.6.4.0-1
This also sets OTD_CONFIGURATIONS when generating rules, as the script
will otherwise probe for git root which is guaranteed to break with most AUR helpers.

Also remove a superfluous message from the .install script.
This is only necessary when udev rule triggers aren't working, which
they should've been for a while now.