Fix license and depend on shared object files
By depending on the shared object files instead of the packages that
provide them, makepkg will insert the soname version into the
dependency, so pacman will prodoce an error when trying to update to
incompatible library versions.

Also pnc was relicensed to gplv3 when it was forked from pn. I forgot to
update this when I copied the pkgbuild, so fix it now.