1.24.0
Hopefully this makes the people who are trying to purge this package
from AUR happy. I could not get time for this due to some circumstances,
and will not be able to once again update this for some upcoming
months, so kindly do not send removal requests for this package for the
time being.