kdumpst: Bump to v1.1
So, this is the version v1.1 - mainly it resolves some issues with the
minimal initramfs creation (including a lame bug - I forgot a "-k" on
mkinitcpio, duh!).

Regarding features, 2 main ones:
a) We now support loading config files from /etc - yay!

b) The tool can use "crashkernel" area now to hold a ramoops pstore
registry, if that's the user's wish (not a default).

Finally, special thanks for the Arch Linux user/maintainer Robin Candau
(Antiz) for a good idea that helped a lot with our alpm hooks =)

Signed-off-by: Guilherme G. Piccoli <gpiccoli@igalia.com>