Fix 'externally-managed-environment' and 'no f2py3.11/f2py3' errors
For 'externally-managed-environment', the fix is to remove "$pkgdir"
from the package only at the end, after activating and running pip
install. This means that python-pip and python-setuptools are not
required, since they are installed in the virtual environment and those
will be used.

For 'f2py3: No such file or directory' and similar errors, the fix is to
avoid a fixed list of files to remove "$pkgdir" from, and to remove this
path from all non-binary files in the package.

Limitation: The sed command to remove "$pkgdir" would still fail if the
build directory path contains ',' or regex reserved characters.