Add support for toggling ALL_LIBCMD with a variable
Set ${_all_libcmd} to '1' or 'yes' to enable the full set of ksh
builtins available in libcmd. These are listed via the 'builtin'
built-in and can be executed via the /opt/ast/bin virtual file path or
by loading them with builtin(1).

This also updates the fallback to use -O2 for better performance
(although in normal circumstances it's never used).