Undo e999aac; keep CARCH in executable filenames.
This is important for makepkg -g, as otherwise it will think that the
x86_64 executable is also the armv7h and aarch64 executables. This
results in the produced checksum being incorrect for all architectures
but the first downloaded.

The only difference in this commit compared to how things used to be
before e999aac is that the CARCH variable is no longer used on the
source lines, as that will cause all executable names to use only one
architecture: the one that is targeted when running makepkg. If you look
at the .SRCINFO before e999aac, you can see that x86_64 is in all
filenames, even the armv7h and aarch64 ones. The CARCH variable is still
used when running `install`, however, as in that case you actually want
the executable for whichever architecture is being targeted.