build: python-setuptools is actually a dependency
python-setuptools was classified as a makedependency, which implies that
it should be possible to remove python-setuptools after building the
package and the software would still work.

However, this is not the case since python-setuptools provides the
pkg_resources package which is later imported by the /usr/bin/tootstream
script, and cleaning makedependencies will render tootstream unable to
open.

This commit moves python-setuptools to the dependents section so that it
survives cleaning orphan dependencies on a system if no other package
depends on python-setuptools already.