Bumped version to 7.8.1
jEveAssets 7.8.1
----------------
Bug Fixes:
-Removed outbid auto-update due to impact on esi
-Added advanced contracting to the slots tool (Reported by MagnarM)

Statement regarding removing features:
jEveAssets outbid auto update had a negative impact on ESI (and thereby Eve) - that is not acceptable! ESI is a shared resource and that means everyone who use it have to use it responsible. In this case the means I had to remove the functionality from jEveAssets, which isn't something I like doing, but, it had to be done to be a good esi citizen, which is absolutely required.