upgpkg: 1.6.1-2
Revert back to source archive - tests not passing was a
issue/peculiarity on my end. My BUILDDIR was located in a Git repository
with no commits, a rare setup to say the least. This caused the 'go
build' invocation in internal/cloud/e2e/main_test.go to fail.

Additionally, add a prepare function downloading dependencies to avoid
fetching content in the build function.