Bump to tootstream 0.4.0
Bumps PKGBUILD to 0.4.0, the latest released and numbered version of
this package.

A patch is provided to fix the runtime errors introduced by a breaking
change in an upstream dependency. Patch is taken from commit
32041df30c838a75bc2863a36eb1442bfaba3e70 from the tootstream repository.
(See https://github.com/magicalraccoon/tootstream/issues/231). Aside
from that, no other commits are cherry-picked (that's the purpose of
tootstream-git, which is the Git tip).

The package currently compiles, sometimes shows a warning about using a
deprecated setup.py build command, but these are part of the build
instructions of the upstream repository so I am not touching that.

Also updates SRCINFO and moves former maintainers as contributors.