Fix issues with adding an image / url to vision model requests.
A few weeks ago, on Ollama integration support landing, the _mediachatf() function was changed, and some final fixes are needed to have it working with all the functionality that is intended.

Fix an issue that the script would repeatedely make new requests on the /regen command when there was a bad response.

Chat command //regen was added, it matches the /rr command (edit + regen).

Improve Googl AI integration.

Google token count function will print an "o" while it its running (because it takes a little time for Google to answer, and the script is not to blame for these slwo requests).

Improve the command /regen ("/" and "//" commands, too), and its integration with the text editor interface.

Homogenise code leading whitespaces/tabs.

The fallback _unescapef() command "printf %b" does not unescape '\"', so the fallback function must not be used with the chat /grep command... Anyways, the jq command for unescaping does not seem to be much slower!