fix pkgver and switch back to minetest_game ID
The `minetest-common` package fortunately stopped packing
`minetest_game`, this AUR package now can ship Minetest Game using the
original and unmodified ID.

Minetest Game in the CDB switched to a different title (instead of a
version number they now use the date). For being independent of the used
title, this package now uses the release date of the latest CDB package
when building the Arch package.

@see https://gitlab.archlinux.org/archlinux/packaging/packages/minetest/-/commit/7d8c7e5100028ff0fcba106d4aea67871526a5dd

@see https://content.minetest.net/packages/Minetest/minetest_game/releases/