Switch to building from source tarball and make package reproducible
Switch to source tarball since we do not need git for building.
Adjust sources for test assumptions in prepare() and not in check().
Provide date to the build process with SOURCE_DATE_EPOCH in mind to make
it reproducible.
Provide a dummy commit checksum to the build as we build from source
tarball and do not care about this information.
Disable LTO build as it currently renders GO BUILDID non-reproducible.

Signed-off-by: David Runge <dvzrv@archlinux.org>