Fix failing patches when using `makepkg` directly
Typically, `${srcdir}` is below the AUR worktree
when building the package. In that case, the
`git apply` command skips patching but still
returns an exit code of 0, causing patches to be
silently skipped.

This may not happen when using AUR helpers such as
aurutils (`aur build -c`) to build inside a clean
chroot. But in the common case (using `makepkg`
directly), it does.

Fix the issue by temporarily creating a Git
repository inside the extracted source tree before
running `git apply`.

Thanks to users @Henry-ZHR and @Potajito for their
bug reports and guidance.