First steamtokendumper-bin PKGBUILD
The configuration file is now readable by the binary but since the
directory isn’t world-writable (and shouldn’t be!) it doesn’t allow for
storing Steam credentials etc. in it.

There is no upstream versioning that I can find and downloads seem to
use the (SHA-256) hash provided at https://steamdb.info/tokendumper/ so
using a $pkgver scheme similar to that of VCS packages[1] prefixed with
a "0." allowing me to bump versions without bumping $epoch.
[1] https://wiki.archlinux.org/title/VCS_package_guidelines#Git

I might look into using unionfs-fuse as suggested for Wine packages[2]
to see if that could provide a viable work-around for this.
[2] https://wiki.archlinux.org/title/Wine_package_guidelines#UnionFsFuse