Change pkgver to avoid '+' in commit hash
Instead use the `.g` prefix for commit hashes
similar to VCS packages: '0.12.0_dev.3433.g640acf862'

Although this commit changes the versioning scheme,
changing the epoch should not be needed since the format of the
primary version number and revision count remains the same.

This fixes an issue with '+' being a special character in URLs.

The build script now requires 'ripgrep' in order to rewrite versions.
I could have used python or jq but this is much more readable.