labwc-atomic-tearing-git: Remove environment variable check
Also change dependencies to a version of wlroots that will include the requisite
patch to support atomic modesetting tearing.

This package and fork are temporary, until upstream approves this change, which
will apparently hinges on guarantees that turning it on won't simply cause a
lot of log spam, which is impossible if the hardware itself doesn't support it,
until such time that wlroots can report to the compositor that support is
present.