Replace prebuilt .tar.xz with install script
'Orchis.tar.xz' is not built automatically every release and tends to be outdated in
the upstream repository. The solution for this to build the package from source,
so we guarantee the installed version is correct.

Note that build the default theme doesn't require sassc and uses prebuild .css files.