Improve text editor interface, and the continue option "-C".
Check and set a different text buffer file name for the editor function when the buffer file is open with a text editor.

Remove the last line of the history file if that contains the "SESSION BREAK" tag before initiating a new session. This improves setting the option -C (continue from last session) when a chat was initiated but exited without actually recording anythinng new. So then, the -C (continue) option will really continued from the last session!

Add an empty line to stderr before printing the text editor buffer in the confirmation prompt.