Update to build again
At lot has changed since the initial package creation. Qt 6 is used now,
`libqatemcontrol` is no longer used and building got significantly
simpler, requiring just the build script to be invoked.

The only submodule is now also tracked as suggested in the
documentation[^1].

The output of the build script is a .deb file which needs to be
extracted to obtain all files that are supposed to be in the package.

[^1]: https://wiki.archlinux.org/title/VCS_package_guidelines#Git_submodules