Package CasparCG client 2.3.0
The binary made available by upstream is for Ubuntu 22 and depends on
Boost 1.74. This version could be packaged in the AUR but possibly more
problems might arise after that.

Instead this package takes the source code of the CasparCG client and
builds it. Since building the client has become significantly simpler
this was preferred over trying to force the binary to work.

The source code is obtained via git instead of the source code archive
associated with the release on GitHub because that archive lacks the
required submodule for building.