Fix "chat command /url" to dump text from an URL.
The function would not work correctly because we mistakenly used echo commands inside set_browsercmdf(). Now it is using printf with no ending newline.

This error was made soon after first implementing the new /url chat command, so it has been there for a couple of months, since the command inception (Feb 8th).