Apply some patches of the Text-PDF bug-tracker.
The bug-tracker can be found at
https://rt.cpan.org/Public/Dist/Display.html?Name=Text-PDF
The number in the patch-filename refers to the specific bug-id there.

Note: I first apply the patch of bug-id 124350 and afterwards bug-id 123562.
So the changes that were already applied from the first patch
are not a part of the second one.
So the patch in the bug-tracker differs accordingly
from bug123562.patch provided in the package.