update to most recent version
In addition to the update, the date (version number) is now show in the
client as addition to the game title:

    Minetest Game (CDB)  ->  Minetest Game (CDB, VERSION)

Since the authors use the current date in the format `YYYYMMDD` as
version number, it will be the date of the release of that version.

The game ID is not changed, not is anything else except the title
and the icon changed or altered in any way.