Limit failed image/url loading for multimodel model.
In multimodal models, such as vision models, it would try to check through the entire input prompt for an image/url to upload to the model when there was a pipe "|" character at the end of the prompt. Such is the expected behaviour,because we decided to be forgiving while trying to load images/urls, but as per my own experience, it should be limited to avoid unecessary printing out of errors when the input contains pipe characters but no image/url to load.