Improve how $REPLY_OLD is set. *Unescape the function json output*.
If the Continue option -C is set, we will set $REPLY_OLD from the last Question recorded in the history file. This improves using the regenerate response functions when running the script on different history files (sessions from different history files).

Note that we will load the the last Question from the history file even if that is a killed line (starts with a comment "#")..  This should be a little better than grepping `grep -ve "^#"` to get the last uncommented Question entry, I reckon, for some user interface behaviours on regenerate...