0.5.4-3: Change build from setuptools to pip
The "python setup.py install" method is deprecated [1] and replaced by
"pip install ."

However there's a bug[2] in pip since ~9 years which necessitates a
hack so it does not try to uninstall the globally installed version.
This bug triggers only when building on a system with this package
previously installed with the _same_ pkgver (i.e. when rebuilding, or
when the pkgrel is increased).

[1] https://packaging.python.org/en/latest/discussions/setup-py-deprecated/
[2] https://github.com/pypa/pip/issues/3063
    "install --root=... tries to remove previously globally installed version"