Version 0.7.0
See the Idris2 CHANGELOG for a full list of changes.

This has been out for a while but the PKGBUILD was broken (and most of
Idris2 users probably use `pack` these days). I believe I have fixed the
PKGBUILD (it seems the build process is a bit better behaved these days)
although this will become apparent once more people install it.