Minor adjustment in TTS: initial buffering timeout is now 2 secs (was 3 secs)
Actually, now the minimum amount of time if 2 secs but the buffering may go on to as almost 3 seconds.

Function new_prompt_confirmf() receives improved case statements to check and set up for extra options.

New "option whisper_append" as "[W]hspr_Add", to append more input to the previous input in whisper mode.
Type "W" or "A" to activate it (only for the full verbose interface).

Improve the chat command "-w" (or "/w") to accept one-shot whisper mode with "-ww" and "/ww".

OBS: case statements are used because they are faster than [[ test for glob matching.

<!-- cancelled: Option "redo" in text editor interface erases input, as it did a long time ago but became the same as the "edit option", I dont remember why. -->