Regression: fix session management commands
Session management commands (`!session`, `!copy`, and `!fork`),
were not working correctly since at least commit
7b94eab386185cefb5286727e8590582007c9f5e (at gitlab)
on 15 May 2024 at 19:04:30 GMT-3 (3 weeks ago).

This is due to how the history recording system has changed since the start of this year,
which in retrospective I did not think this refactoring would change almost all
functions dealing with session and history!

I reckon that the change was good because it stopped littering the history file
with unecessary buffers but breaking a session now has become much of a larger problem.

Also, add some more examples of how these commands may be used to copy an older session to the current (tail) session of the history file. Not so straightforwards but the simplest I can think of, anyways.