Print session context on -C (continue), aglutinate code of session and chat commands.
The session commands can now be aglutinated with the chat commands becase they dont break the session in a history file before the end  of the loop now!

However, to kep the  old behavior to change to a history file as a positional parameter `/[histfile]` shortcut for `/session [histfile]`, we dont remove the old code.