Optimise start-up time in some circunstances.
Only preview input prompt tokens when that is not a session nor a chat command. For that, we were able to change the code position in the script so that is only activated when we are certain of user input from the command line.

Also, move the block of warning and tips with that.

Revert a change in recognising session commands from the command line input I had dome as an experiment but forgot to remove in the last update.