chore(PKGBUILD): update vimgolf package version
In this commit, the vimgolf package version in the PKGBUILD file has been updated from 0.4.0 to 0.5.1. The md5sums has been updated accordingly to match the new version's checksum. This change ensures that we are using the latest version of the vimgolf package, which may include new features, bug fixes, or performance improvements. As always, the update was straightforward and didn't pose any challenges.