Tidy the code, and improve the chat command "!regen" (regenerate last response).
Update unset variables on some functions that break the session.

Add code into a function grep_usr_lastlinef() to re-use it in some other
specific cases to activate the regeneration mechanism (last response regeneration)
to grep and load the last user entry from history file as the prompt
to make the new request.

I think this solves some issues when changing history files, resuming/
continuing the last session from this other history file,
and trying to regenerate the last response from the last question (request prompt).

This ensures that the last regeneration of response matches the session of the
history file even if it has changed and resumed (so no buffer in the running programme yet).

I use the regenerate functions (mainly the alias "//" on a new prompt) to
regenerate the last response of a session, and I have been experimenting a lot with changing from history files.

The new update makes the chat command "/regen" or its alias "/", or even "//" on a new prompt to correctly regenerate the last response when changing to
different history files.

Also, just improve how some code is spaced and paragraphed to better reading.

Overall, in v0.60 through v0.61, the script has regained code optimisations
for speed, some updated benefited even the start up time, which should be noticeable specially for google gemini models.