Improve retry request on TTS curl error.
Since commit
e2bd85172f944d9017239e2b878d475033072aa2 (gitlab)
v0.24.14  dec/2023
the retry request in TTS curl error functionality seem to have been broken.

We updated the kill command with the "-s INT", so ctrl-c and pressing [k] will have the same effect,
we also wants to be sure to have a non-zero exit code for checking in the following case statement.

And we updated the case statement with older conditions, so that it now works as expected.

Update all curl commands with --fail-with-body and --fail only, because --fail-with-body was implemented in curl
Added in 7.76.0
31-Mar-2021

and we want to make sure this script works with older curl versions.