Improve user feedback on Stop sequences chat command "-s [seq]".
The chat command "-s [seq]" or "!stop [seq]" now displays the feedback of the input sequences more correctly as we escape the stop sequence buffer one by one, similarly how the codes proceeds with the "!info" chat command which also prints the stop sequences.

We also made the leading white space globbing more weak in chat commands "!start [seq]", "!restart [seq]", and "!stop [seq]". This means the code will glob out *only the first white space* after "!stop" command keyword to set the sequence that comes after it, meaning that a sequence of multiple white spaces is now recognised and set as a target.

On "chatgpt.sh -ccvvv", the code will feedback print to the user a "p" instead of an asterix to convey that the TTS function is going to play right away (there is still the 2 second buffer that applies).

Majorly minor improvements, just a little polishing.